Added access checks to projectService

+++ server/src/main/java/eu/openanalytics/phaedra/plateservice/service/ProjectService.java
@@ -1,15 +1,17 @@
 package eu.openanalytics.phaedra.plateservice.service;
 
-import eu.openanalytics.phaedra.plateservice.model.Project;
-import eu.openanalytics.phaedra.plateservice.repository.ProjectRepository;
-import eu.openanalytics.phaedra.platservice.dto.ProjectDTO;
+import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
+
 import org.modelmapper.Conditions;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
+import eu.openanalytics.phaedra.plateservice.model.Project;
+import eu.openanalytics.phaedra.plateservice.repository.ProjectRepository;
+import eu.openanalytics.phaedra.platservice.dto.ProjectDTO;
+import eu.openanalytics.phaedra.platservice.enumartion.ProjectAccessLevel;
 
 @Service
 public class ProjectService {
@@ -19,12 +21,16 @@ public class ProjectService {
 	
 	private final ExperimentService experimentService;
 
-	public ProjectService(ProjectRepository projectRepository, ExperimentService experimentService) {
+	private final ProjectAccessService projectAccessService;
+	
+	public ProjectService(ProjectRepository projectRepository, ExperimentService experimentService, ProjectAccessService projectAccessService) {
 		this.projectRepository = projectRepository;
 		this.experimentService = experimentService;
+		this.projectAccessService = projectAccessService;
 	}
 
 	public ProjectDTO createProject(ProjectDTO projectDTO) {
+		projectAccessService.checkCanCreateProjects();
 		Project project = new Project();
 		modelMapper.typeMap(ProjectDTO.class, Project.class)
 				.map(projectDTO, project);
@@ -35,6 +41,7 @@ public class ProjectService {
 	public void updateProject(ProjectDTO projectDTO) {
 		Optional<Project> project = projectRepository.findById(projectDTO.getId());
 		project.ifPresent(p -> {
+			projectAccessService.checkAccessLevel(p.getId(), ProjectAccessLevel.Write);
 			modelMapper.typeMap(ProjectDTO.class, Project.class)
 					.setPropertyCondition(Conditions.isNotNull())
 					.map(projectDTO, p);
@@ -43,6 +50,7 @@ public class ProjectService {
 	}
 
 	public void deleteProject(long projectId) {
+		projectAccessService.checkAccessLevel(projectId, ProjectAccessLevel.Admin);
 //		experimentService.deleteExperimentsByProjectId(projectId);
 		projectRepository.deleteById(projectId);
 	}
@@ -50,13 +58,17 @@ public class ProjectService {
 	public List<ProjectDTO> getAllProjects() {
 		List<Project> projects = (List<Project>) projectRepository.findAll();
 		return projects.stream()
+				.filter(p -> projectAccessService.hasAccessLevel(p.getId(), ProjectAccessLevel.Read))
 				.map(this::mapToProjectDTO)
 				.collect(Collectors.toList());
 	}
 	
 	public ProjectDTO getProjectById(long projectId) {
 		Optional<Project> result = projectRepository.findById(projectId);
-		return result.map(this::mapToProjectDTO).orElse(null);
+		return result
+				.map(this::mapToProjectDTO)
+				.filter(p -> projectAccessService.hasAccessLevel(p.getId(), ProjectAccessLevel.Read))
+				.orElse(null);
 	}
 
 	private ProjectDTO mapToProjectDTO(Project project) {
